pyproject.toml
[project]
name = "intersect"
version = "0.1.0"
description = "Job search and analysis tool with DuckDB storage, semantic search, and NLP capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "bm25s>=0.2.6",
    "cohere>=5.13.6",
    "curl-cffi>=0.7.4",
    "duckdb>=1.3.2",
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "openai>=1.59.3",
    "pandas>=2.2.3",
    "pypdf>=5.1.0",
    "pystemmer>=3.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "scikit-learn>=1.6.0",
    "selectolax>=0.3.27",
    "spacy>=3.8.3",
    "streamlit>=1.47.1",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
    "wordcloud>=1.9.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
